move tests to jest
