+++ src/utils.ts
@@ -87,41 +87,3 @@ export function flatMap<A, B>(xs: A[], fn: (a: A) => B[]): B[] {
 }
 
 export type AllOrNothing<T> = T | { [key in keyof T]?: never };
-
-// lots of ts hackery to unit test a type hehe
-// @ts-ignore
-// eslint-disable-next-line @typescript-eslint/no-unused-vars
-namespace Tests {
-  // @ts-ignore
-  type Extends<A, B> = B extends A ? 'true' : 'false';
-  type AssertTrue<A extends 'true'> = Extends<'true', A>;
-  type AssertFalse<A extends 'false'> = Extends<'false', A>;
-  type AllTrue<A extends 'true'[]> = Extends<'true'[], A>;
-
-  type Person = { name: string; age: number };
-
-  type test_AllOrNothing_accepts_all = AssertTrue<
-    Extends<AllOrNothing<Person>, { name: 'Joe'; age: 100 }>
-  >;
-
-  type test_AllOrNothing_does_not_accept_partial = AssertFalse<
-    Extends<AllOrNothing<Person>, { name: 'joe' }>
-  >;
-
-  type test_AllOrNothing_accepts_nothing = AssertTrue<
-    Extends<AllOrNothing<Person>, {}>
-  >;
-
-  type test_AllOrNothing = AssertTrue<
-    AllTrue<
-      [
-        test_AllOrNothing_accepts_all,
-        test_AllOrNothing_does_not_accept_partial,
-        test_AllOrNothing_accepts_nothing
-      ]
-    >
-  >;
-
-  // @ts-ignore
-  type all_tests = [test_AllOrNothing];
-}
