+++ test/utils.test.ts
@@ -1,4 +1,4 @@
-import { padNoAnsi } from '../src/utils';
+import { padNoAnsi, AllOrNothing } from '../src/utils';
 import stripAnsi from 'strip-ansi';
 import chalk from 'chalk';
 
@@ -27,3 +27,38 @@ describe('padNoAnsi', () => {
     expect(actual).toEqual(str);
   });
 });
+
+it('passes type tests', () => {
+  function identity<T>(t: T): T {
+    return t;
+  }
+
+  expect(identity<TypeTests.test>('true')).toEqual('true');
+});
+
+namespace TypeTests {
+  type Extends<A, B> = B extends A ? 'true' : 'false';
+  type AssertTrue<A extends 'true'> = Extends<'true', A>;
+  type AssertFalse<A extends 'false'> = Extends<'false', A>;
+  type AllTrue<A extends 'true'[]> = Extends<'true'[], A>;
+
+  namespace AllOrNothingTests {
+    type Person = { name: string; age: number };
+
+    type accepts_all_elements = AssertTrue<
+      Extends<AllOrNothing<Person>, { name: 'Joe'; age: 100 }>
+    >;
+
+    type does_not_accept_partial = AssertFalse<
+      Extends<AllOrNothing<Person>, { name: 'joe' }>
+    >;
+
+    type accepts_nothing = AssertTrue<Extends<AllOrNothing<Person>, {}>>;
+
+    export type test = AssertTrue<
+      AllTrue<[accepts_all_elements, does_not_accept_partial, accepts_nothing]>
+    >;
+  }
+
+  export type test = AllTrue<[AllOrNothingTests.test]>;
+}
